ライン破壊アニメーションを実装

+++ game.cpp
@@ -1,15 +1,21 @@
 #include "game.h"
 #define DEFAULT_TURN_LENGTH 30
+#define DELETE_ANIMETION_LENGTH (DEFAULT_TURN_LENGTH/2) 
 
 Game::Game(){
   _keys = new Keys();
   _field = new Field();
   _mino = new Mino(_keys, _field);
   _turn_counter = 0;
+  _delete_animation_counter = 0;
   _turn_length = DEFAULT_TURN_LENGTH;
 }
 
 void Game::update(){
+  if(_delete_animation_counter > 0){
+    updateAsDeleteAnimation();
+    return;
+  }
   _mino->update();
   _keys->reset();
   if(_turn_counter >= _turn_length){
@@ -20,10 +26,17 @@ void Game::update(){
   _turn_counter++;
 }
 
+void Game::updateAsDeleteAnimation(){
+  bool visible = _delete_animation_counter / (DELETE_ANIMETION_LENGTH / 4) % 2;
+  _field->updateAsDeleteAnimation(visible);
+  _delete_animation_counter--;
+  if(_delete_animation_counter == 0) _field->deleteLines();
+}
+
 void Game::putMino(){
   _mino->fix();
   if(_field->deleteTest()){
-    _field->deleteLines();
+    _delete_animation_counter = DELETE_ANIMETION_LENGTH;
   }
   _mino->refresh();
 }
